Feat: 카카오 로그인 변수 수정

+++ src/main/java/com/project/final_project/login/kakao/service/KakaoService.java
@@ -117,7 +117,7 @@ public class KakaoService {
     JSONObject account = (JSONObject) jsonObj.get("kakao_account");
     JSONObject profile = (JSONObject) account.get("profile");
 
-    long id = (long) jsonObj.get("id");
+    Integer id = (Integer) jsonObj.get("id");
     String email = String.valueOf(account.get("email"));
     String nickname = String.valueOf(profile.get("nickname"));
